changed to double

+++ src/main/java/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int result = 10/3;
+        double result = (double)10 / (double)3;
         System.out.println(result);
     }
 }
